+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All significant changes to this project will be documented here.
 
 ## [Unreleased]
 
-*   Make `_auto_from_value` public as `ff.auto_from_value`.
+*   Made `_auto_from_value` public as `ff.auto_from_value`.
+*   Dropped Python 3.8 support.
 
 ## [1.2]